requirements.txt
langchain
langchain-community
sentence-transformers
docling
langchain_groq
langchain-chroma
chromadb
rank-bm25
python-dotenv
chainlit
scikit-learn